upgraded n_consumed to long long

+++ lib/hashtable.cc
@@ -315,7 +315,7 @@ unsigned int Hashtable::check_and_process_read(const std::string &read,
 
 void Hashtable::consume_fasta(const std::string &filename,
 			      unsigned int &total_reads,
-			      unsigned int &n_consumed,
+			      unsigned long long &n_consumed,
 			      HashIntoType lower_bound,
 			      HashIntoType upper_bound,
 			      ReadMaskTable ** orig_readmask,
